Split `migrate-1x` into its own standalone `dbt-migrate-1x` package
Moves the command out of the `dbt-autofix` CLI into a separate uv workspace
package so it can be invoked directly (e.g. by the dbt VS Code extension)
without going through the main CLI.
