+++ packages/dbt_migrate_1x/.gitignore
@@ -1,0 +1,14 @@
+# Python-generated files
+__pycache__/
+*.py[oc]
+build/
+dist/
+wheels/
+*.egg-info
+
+# Virtual environments
+.venv
+.nox
+venv
+
+.DS_Store
