+++ packages/dbt_migrate_1x/.python-version
@@ -1,0 +1,1 @@
+3.10
