+++ packages/dbt_migrate_1x/pyproject.toml
@@ -1,0 +1,36 @@
+[project]
+name = "dbt-migrate-1x"
+description = "Deterministic dbt 1.x-to-1.x migration CLI, runnable directly via uv"
+readme = "README.md"
+requires-python = ">=3.10, <3.14" # can't support 3.14 without mashumaro 3.15
+dependencies = [
+    "dbt-autofix",
+    "typer>=0.16.0",
+    "rich>=13.7.0",
+]
+
+dynamic = ["version"]
+
+[project.scripts]
+dbt-migrate-1x = "dbt_migrate_1x.cli:app"
+
+[tool.uv]
+package = true
+
+[tool.uv.sources]
+dbt-autofix = { workspace = true }
+
+[build-system]
+requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
+build-backend = "hatchling.build"
+
+[tool.hatch.version]
+source = "uv-dynamic-versioning"
+
+[tool.uv-dynamic-versioning]
+vcs = "git"
+style = "pep440"
+# Fallback when git tags are unavailable (e.g. Dependabot running uv lock in a shallow clone)
+fallback-version = "0.0.0"
+# Don't include the commit ID in the version - required for TestPypi
+metadata = false
