+++ pyproject.toml
@@ -14,6 +14,7 @@ cache-keys = [{ file = "pyproject.toml" }, { git = { commit = true, tags = true 
 [tool.uv.workspace]
 members = [
     "packages/dbt_fusion_package_tools",
+    "packages/dbt_migrate_1x",
 ]
 
 [tool.uv.sources]
